plcbrowserapi/requirements.txt
fastapi==0.85.1
python-dotenv==0.21.0
uvicorn==0.19.0
pycomm3==1.2.10